refactor(bridge): extract presence lifecycle
